requirements.txt
pyrogram==2.0.106
py-tgcalls==2.0.6
flask==3.0.3
flask-limiter==3.8.0
bcrypt==4.1.3
aiosqlite==0.20.0
yt-dlp==2024.5.27
python-dotenv==1.0.1
